+++ art/__init__.py
@@ -1,1 +1,2 @@
+"""Art modules."""
 from .art import *