+++ art/art.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Art module."""
 from .art_dic import *
 from .art_param import *
 import os
@@ -7,21 +8,25 @@ import random
 
 
 class artError(Exception):  # pragma: no cover
+    """Art error class."""
+
     pass
 
 
 def font_size_splitter(font_map):
-    '''
-    This function split fonts to 4 category (small,medium,large,xlarge) by maximum length of letter in each font
+    """
+    Split fonts to 4 category (small,medium,large,xlarge) by maximum length of letter in each font.
+
     :param font_map: input fontmap
     :type font_map : dict
     :return: splitted fonts as dict
-    '''
+    """
     small_font = []
     medium_font = []
     large_font = []
     xlarge_font = []
-    for font in font_map.keys():
+    fonts = set(font_map.keys()) - set(RANDOM_FILTERED_FONTS)
+    for font in fonts:
         length = max(map(len, font_map[font][0].values()))
         if length <= FONT_SMALL_THRESHOLD:
             small_font.append(font)
@@ -42,47 +47,39 @@ RND_SIZE_DICT = font_size_splitter(FONT_MAP)  # pragma: no cover
 
 
 def line(char="*", number=30):
-    '''
-    This function print line of chars
+    """
+    Print line of chars.
+
     :param char: input character
     :type char:str
     :param number: number of characters
     :return: None
-    '''
+    """
     print(char * number)
 
 
-def font_list(text="test", test_mode=False):
-    '''
-    This function print all Of fonts
+def font_list(text="test"):
+    """
+    Print all fonts.
+
     :param text : input text
     :type text : str
-    :param test_mode : test mode activation flag
-    :type test_mode : bool
     :return: None
-    '''
+    """
     for item in sorted(list(FONT_MAP.keys())):
         print(str(item) + " : ")
         text_temp = text
-        if test_mode:
-            if str(item) in [
-                "char4",
-                "c2",
-                "war_of_w",
-                "coil_cop",
-                "fbr12",
-                    "ghost_bo"]:
-                text_temp = text_temp.upper()
         tprint(text_temp, str(item))
 
 
 def art_list(test=False):
-    '''
-    This function print all Of 1-Line arts
+    """
+    Print all 1-Line arts.
+
     :param test : exception test flag
     :type test : bool
     :return: None
-    '''
+    """
     for i in sorted(list(art_dic.keys())):
         try:
             if test:
@@ -98,10 +95,11 @@ def art_list(test=False):
 
 
 def help_func():
-    '''
-    Print help page
+    """
+    Print help page.
+
     :return: None
-    '''
+    """
     tprint("art")
     tprint("v" + VERSION)
     print(DESCRIPTION + "\n")
@@ -117,23 +115,24 @@ def help_func():
 
 
 def aprint(artname, number=1, text=""):
-    '''
-    Art print
-    This function print ascii art
+    """
+    Print 1-line art.
+
     :param artname: artname
     :type artname : str
     :return: None
-    '''
+    """
     print(art(artname=artname, number=number, text=text))
 
 
 def art(artname, number=1, text=""):
-    '''
-    This function return ascii art
+    """
+    Return 1-line art.
+
     :param artname: artname
     :type artname : str
     :return: ascii art as str
-    '''
+    """
     if isinstance(artname, str) is False:
         raise artError("artname shoud have str type")
     artname = artname.lower()
@@ -161,16 +160,18 @@ def art(artname, number=1, text=""):
 
 
 def randart():
-    '''
-    This function return random 1-line art
+    """
+    Return random 1-line art.
+
     :return: ascii art as str
-    '''
+    """
     return art("random")
 
 
 def tprint(text, font=DEFAULT_FONT, chr_ignore=True):
-    '''
-    This function split function by \n then call text2art function
+    r"""
+    Split text by \n then call text2art function.
+
     :param text: input text
     :type text:str
     :param font: input font
@@ -178,7 +179,7 @@ def tprint(text, font=DEFAULT_FONT, chr_ignore=True):
     :param chr_ignore: ignore not supported character
     :type chr_ignore:bool
     :return: None
-    '''
+    """
     if isinstance(text, str) is False:
         raise artError("text should have str type")
     split_list = text.split("\n")
@@ -195,8 +196,9 @@ def tsave(
         filename="art",
         chr_ignore=True,
         print_status=True):
-    '''
-    This function save ascii art
+    """
+    Save ascii art.
+
     :param text: input text
     :param font: input font
     :type font:str
@@ -208,7 +210,7 @@ def tsave(
     :param print_status : save message print flag
     :type print_status:bool
     :return: None
-    '''
+    """
     try:
         split_list = text.split("\n")
         files_list = os.listdir(os.getcwd())
@@ -241,8 +243,9 @@ def tsave(
 
 
 def distance_calc(s1, s2):
-    '''
-    This function calculate Levenshtein distance between two words
+    """
+    Calculate Levenshtein distance between two words.
+
     :param s1: first word
     :type s1 : str
     :param s2: second word
@@ -252,7 +255,7 @@ def distance_calc(s1, s2):
     References :
     1- https://stackoverflow.com/questions/2460177/edit-distance-in-python
     2- https://en.wikipedia.org/wiki/Levenshtein_distance
-    '''
+    """
     if len(s1) > len(s2):
         s1, s2 = s2, s1
 
@@ -270,12 +273,13 @@ def distance_calc(s1, s2):
 
 
 def wizard_font(text):
-    '''
-    This function check input text length for wizard mode
+    """
+    Check input text length for wizard mode.
+
     :param text: input text
     :type text:str
     :return: font as str
-    '''
+    """
     text_length = len(text)
     if text_length <= TEXT_XLARGE_THRESHOLD:
         font = random.choice(XLARGE_WIZARD_FONT)
@@ -289,8 +293,9 @@ def wizard_font(text):
 
 
 def indirect_font(font, fonts, text):
-    '''
-    This function check input font for indirect modes
+    """
+    Check input font for indirect modes.
+
     :param font: input font
     :type font : str
     :param fonts: fonts list
@@ -298,7 +303,7 @@ def indirect_font(font, fonts, text):
     :param text: input text
     :type text:str
     :return: font as str
-    '''
+    """
     if font == "rnd-small" or font == "random-small" or font == "rand-small":
         font = random.choice(RND_SIZE_DICT["small_list"])
         return font
@@ -324,8 +329,9 @@ def indirect_font(font, fonts, text):
 
 
 def text2art(text, font=DEFAULT_FONT, chr_ignore=True):
-    '''
-    This function print art text
+    """
+    Return art text.
+
     :param text: input text
     :type text:str
     :param font: input font
@@ -333,7 +339,7 @@ def text2art(text, font=DEFAULT_FONT, chr_ignore=True):
     :param chr_ignore: ignore not supported character
     :type chr_ignore:bool
     :return: ascii art text as str
-    '''
+    """
     split_list = []
     result_list = []
     letters = standard_dic
@@ -349,6 +355,8 @@ def text2art(text, font=DEFAULT_FONT, chr_ignore=True):
     letters = FONT_MAP[font][0]
     if FONT_MAP[font][1]:
         text_temp = text.lower()
+    if font in UPPERCASE_FONTS:
+        text_temp = text.upper()
     for i in text_temp:
         if (ord(i) == 9) or (ord(i) == 32 and font == "block"):
             continue
@@ -379,8 +387,9 @@ def text2art(text, font=DEFAULT_FONT, chr_ignore=True):
 
 def set_default(font=DEFAULT_FONT, chr_ignore=True, filename="art",
                 print_status=True):
-    '''
-    This function change text2art, tprint and tsave default values
+    """
+    Change text2art, tprint and tsave default values.
+
     :param font: input font
     :type font:str
     :param chr_ignore: ignore not supported character
@@ -390,7 +399,7 @@ def set_default(font=DEFAULT_FONT, chr_ignore=True, filename="art",
     :param print_status : save message print flag (only tsave)
     :type print_status:bool
     :return: None
-    '''
+    """
     if isinstance(font, str) is False:
         raise artError("font should have str type")
     if isinstance(chr_ignore, bool) is False: