+++ art/art_param.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
+"""Art parameters."""
 from .text_dic1 import *
 from .text_dic2 import *
 from .art_dic import *
 
-VERSION = "3.1"  # pragma: no cover
+VERSION = "3.2"  # pragma: no cover
 FONT_SMALL_THRESHOLD = 50  # pragma: no cover
 FONT_MEDIUM_THRESHOLD = 100  # pragma: no cover
 FONT_LARGE_THRESHOLD = 200  # pragma: no cover
@@ -50,6 +51,92 @@ XLARGE_WIZARD_FONT = [
     "nscript",
     "georgia11",
     "block"]  # pragma: no cover
+
+UPPERCASE_FONTS = [
+    "1943",
+    "4x4_offr",
+    "64f1",
+    "a_zooloo",
+    "advenger",
+    "aquaplan",
+    "assalt_m",
+    "asslt_m",
+    "atc",
+    "atc_gran",
+    "battle_s",
+    "battlesh",
+    "baz_bil",
+    "beer_pub",
+    "c1",
+    "c2",
+    "c_consen",
+    "char1",
+    "char4",
+    "coil_cop",
+    "druid",
+    "faces_of",
+    "fair_mea",
+    "fairligh",
+    "fantasy",
+    "fbr1",
+    "fbr12",
+    "fbr_stri",
+    "fbr_tilt",
+    "finalass",
+    "fp1",
+    "fp2",
+    "funky_dr",
+    "future_1",
+    "future_2",
+    "future_3",
+    "future_4",
+    "future_5",
+    "future_6",
+    "future_7",
+    "future_8",
+    "ghost_bo",
+    "grand_pr",
+    "green_be",
+    "hades",
+    "heavy_me",
+    "heroboti",
+    "house_of",
+    "hypa_bal",
+    "hyper",
+    "kgames_i",
+    "kik_star",
+    "krak_out",
+    "tsn_base",
+    "ugalympi",
+    "unarmed",
+    "usa",
+    "usa_pq",
+    "vortron",
+    "war_of_w",
+    "yie-ar",
+    "yie_ar_k",
+    "z-pilot",
+    "zig_zag",
+    "zone7"]
+
+RANDOM_FILTERED_FONTS = [
+    "5x8",
+    "binary",
+    "decimal",
+    "high_noo",
+    "hills",
+    "katakana",
+    "morse",
+    "morse2",
+    "nfi1",
+    "octal",
+    "rot13",
+    "smtengwar",
+    "tengwar",
+    "tsalagi",
+    "gauntlet",
+    "flyn_sh",
+    "moscow"]
 
 DESCRIPTION = '''ASCII art is also known as "computer text art".
 It involves the smart placement of typed special characters or
@@ -277,7 +364,7 @@ FONT_MAP = {"block": [block_dic, True], "banner": [banner_dic, False],  # pragma
             "heroboti": [heroboti_dic, False],
             "high_noo": [high_noo_dic, False],
             "hills": [hills_dic, False],
-            "home_pak": [home_pak_dic, False],
+            "home_pak": [home_pak_dic, True],
             "house_of": [house_of_dic, False],
             "hypa_bal": [hypa_bal_dic, False],
             "hyper": [hyper_dic, False],
@@ -345,7 +432,7 @@ FONT_MAP = {"block": [block_dic, True], "banner": [banner_dic, False],  # pragma
             "coil_cop": [coil_cop_dic, False],
             "com_sen": [com_sen_dic, False],
             "druid": [druid_dic, False],
-            "e_fist": [e_fist_dic, False],
+            "e_fist": [e_fist_dic, True],
             "ebbs_1": [ebbs_1_dic, False],
             "ebbs_2": [ebbs_2_dic, False],
             "eca": [eca_dic, False],