+++ art_profile.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-
+"""Profile file."""
 from art import *
 
 font_list("ART PROFILE")