+++ otherfile/test.sh
@@ -3,6 +3,10 @@ set -x
 
 python -m art testcov
 python version_check.py
-python -m vulture --min-confidence 80 --exclude=art,build,.eggs --sort-by-size .
-python -m bandit -r art -s B311
+if [ "$TRAVIS_PYTHON_VERSION" = '3.6' ]
+  then
+      python -m vulture --min-confidence 80 --exclude=art,build,.eggs --sort-by-size .
+      python -m bandit -r art -s B311
+	  python -m pydocstyle --match='(?!test).*\.py'
+  fi
 python -m cProfile -s cumtime art_profile.py